longarr PR #2: fix dasLLVM JIT for widened uint64 Array/Table; format-string and bounds-check follow-ups
Phase 2's struct widening of Array/Table size/capacity to uint64_t broke the
dasLLVM JIT path because its LLVM struct mirror still declared those fields as
i32. Every utility binary built by JIT (dascov/daspkg/dastest/aot/mcp/...) hit
LLVMVerifyFunction with i32-vs-i64 mismatches the moment it touched any
array — which is essentially every script.

Fixed in this commit:
- modules/dasLLVM/daslib/llvm_jit_common.das: widen Array/Table LLVM struct
  mirrors — size, capacity, tombstones now i64 to match arraytype.h layout.
- modules/dasLLVM/daslib/llvm_jit.das: new normalize_int_widths helper
  zero-extends the narrower operand whenever the JIT compares an i32 index
  against an i64 size (default-int indexing into widened arrays). Applied at
  check_range and at the three inline ICmp sites in visitExprSafeAt
  (dim/array/vector). For-loop "first" pseudoiterator over arrays now builds
  zero/one constants matching the size width.
- modules/dasLLVM/daslib/llvm_jit_intrin.das: intrinsic_builtin_length now
  truncates the i64 size back to i32 for the legacy length() return contract.
  TODO Phase 7: emit a panic on size > INT_MAX to match builtin_array_size.

C++ format-string follow-ups (warnings on every platform, errors on stricter
configs once uint64_t flowed through %u):
- src/simulate/simulate_fusion_at_array.cpp (6 macro expansions),
  include/daScript/simulate/aot.h TArray::operator() i32 overloads,
  modules/dasStbImage/src/dasRaster.cpp rast_hspan_*: %u/%i -> %llu with cast.

Copilot review fixes:
- src/builtin/module_builtin_array.cpp builtin_array_erase_range_i64:
  signed overflow UB on (index+count) before the cast - recompute as
  uint64(index)+uint64(count) AFTER non-negativity check.
- daslib/array_boost.das array_view int64 overloads: validate length >= 0
  before the bounds sum, so a negative length doesn't underflow to huge.

Doc stub fixes (das2rst gate was blocking the doc build):
- Wrote handmade .rst entries for long_length/long_capacity (array+table),
  long_iter_range, long_find_index.

+++ include/daScript/simulate/aot.h
@@ -1042,12 +1042,12 @@ namespace das {
     // index
         __forceinline TT & operator () ( int32_t index, Context * __context__ ) {
             uint32_t idx = uint32_t(index);
-            if ( idx>=size ) __context__->throw_error_ex("array index out of range, %u of %u", idx, size);
+            if ( idx>=size ) __context__->throw_error_ex("array index out of range, %u of %llu", idx, (unsigned long long)size);
             return ((TT *)data)[index];
         }
         __forceinline const TT & operator () ( int32_t index, Context * __context__ ) const {
             uint32_t idx = uint32_t(index);
-            if ( idx>=size ) __context__->throw_error_ex("array index out of range, %u of %u", idx, size);
+            if ( idx>=size ) __context__->throw_error_ex("array index out of range, %u of %llu", idx, (unsigned long long)size);
             return ((const TT *)data)[index];
         }
         __forceinline TT & operator () ( uint32_t idx, Context * __context__ ) {
